Rename to Mosh, update readme
It is not a good idea to include a dot in the name: Markdown interprets
it as an URL and GitLab may not be happy with it, it may even cause
problems in other places as well.

The Mosh word is a reflection of what the program is doing: it runs
around the directories.

https://www.youtube.com/watch?v=fgBseiTlDTE